diagonal from int to float32

+++ structs/techniqueStructs/TV.go
@@ -4,6 +4,6 @@ type TV struct {
 	ID        int
 	Technique Techniques
 	Name      string
-	Diagonal  int
+	Diagonal  float32
 	IsSmart   bool
 }
